requirements.txt
google-genai>=1.0.0
tweepy>=4.14.0
pyTelegramBotAPI>=4.21.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=5.0.0
